Refactor Existing tests to better follow code quality standards
- move low lever dom access from tests specs to associated support classes
- replace usage of direct api calls and permit values to use correct support methods.
