+++ permit-ui/cypress/e2e/permits/new-permit-page/new-permit-crud.spec.cy.ts
@@ -102,29 +102,4 @@ describe('Create Permit - CRUD Operations', () => {
     });
   });
 
-  describe('Create Permit Performance', () => {
-    it('should complete permit creation within acceptable time limits', () => {
-      const startTime = Date.now();
-
-      UiActions.visitPermitsPage();
-      UiActions.clickNewPermitButton();
-      UiActions.fillPermitFormFromFixture(PermitFixtureKeys.CREATE_THIS_PERMIT);
-      UiActions.clickSubmitButton();
-
-      // Wait for creation to complete
-      cy.wait(1000);
-      UiActions.navigateToPage('last');
-
-      // Find the last row and verify the permit was created there
-      UiAssertions.verifyFixturePermitInLastRow(
-        PermitFixtureKeys.CREATE_THIS_PERMIT
-      );
-
-      // Verify operation completed within reasonable timeframe (< 5 seconds)
-      cy.then(() => {
-        const duration = Date.now() - startTime;
-        expect(duration).to.be.lessThan(5000);
-      });
-    });
-  });
 });
